+++ src/BigCommerce/Templates/Body_Classes.php
@@ -14,8 +14,10 @@ class Body_Classes {
 	 * @filter body_class
 	 */
 	public function set_body_classes( $classes ) {
+		$classes = array_merge( $classes, $this->get_wp_theme() );
+
 		if ( is_singular( Product::NAME ) ) {
-			$classes = array_merge( $classes, $this->product_single_classes( get_queried_object_id() ) );
+			$classes = array_merge( $classes, $this->product_single_classes( get_queried_object_id() ), $this->get_wp_theme() );
 		}
 
 		return $classes;
@@ -24,7 +26,7 @@ class Body_Classes {
 	private function product_single_classes( $post_id ) {
 		$classes = [];
 		$product = new Product( get_queried_object_id() );
-		
+
 		$classes[] = sprintf( 'bc-product-%d', $product->bc_id() );
 		$classes[] = sprintf( 'bc-availability-%s', $product->availability() );
 
@@ -36,6 +38,17 @@ class Body_Classes {
 		}
 		if ( $product->low_inventory() ) {
 			$classes[] = 'bc-product-lowinventory';
+		}
+
+		return $classes;
+	}
+
+	private function get_wp_theme() {
+		$theme   = wp_get_theme();
+		$classes = [];
+
+		if ( 'Twenty Twenty' === $theme->name || 'Twenty Twenty' === $theme->parent_theme ) {
+			$classes[] = 'bc-wp-twenty-twenty-theme';
 		}
 
 		return $classes;